scenarios/security_arena/agents/normal_user/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
uvicorn>=0.30.0

